requirements.txt
pip~=20.2.3
lxml~=4.6.3
setuptools~=49.2.1
python-dateutil~=2.8.1
jsonschema~=3.2.0
Flask~=1.1.2
pandas~=1.2.3